backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.2",
    "google-cloud-datastore>=2.21.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dateutil>=2.9.0.post0",
    "uvicorn[standard]>=0.38.0",
]
